bring helpers to getD2Manifest
